chore: dataset submissions lambda Docker image cleanup (#2491)
