+++ backend/corpora/dataset_submissions/requirements.txt
@@ -1,10 +1,3 @@
-# TODO: cleanup after authz refactor
-connexion[swagger-ui]==2.7.0
-Flask-Cors>=3.0.6
-python-jose[cryptography]>=3.1.0
 psycopg2-binary>=2.8.5
 SQLAlchemy>=1.3.17,<1.4.0
-Authlib==0.14.3
-rsa>=4.7 # not directly required, pinned by Snyk to avoid a vulnerability
-gunicorn[gevent] >=20.1.0, <21.0.0
 boto3
